+++ src/appleseed/renderer/kernel/shading/closures.h
@@ -35,8 +35,8 @@
 #include "renderer/modeling/bsdf/disneybrdf.h"
 #include "renderer/modeling/bsdf/lambertianbrdf.h"
 #include "renderer/modeling/bsdf/orennayarbrdf.h"
-#include "renderer/modeling/bsdf/oslmicrofacetbrdf.h"
-#include "renderer/modeling/bsdf/oslmicrofacetbtdf.h"
+#include "renderer/modeling/bsdf/osl/oslmicrofacetbrdf.h"
+#include "renderer/modeling/bsdf/osl/oslmicrofacetbtdf.h"
 #include "renderer/modeling/bsdf/specularbrdf.h"
 #include "renderer/modeling/bsdf/specularbtdf.h"
 #include "renderer/modeling/bsdf/velvetbrdf.h"
