+++ src/appleseed/renderer/modeling/bsdf/osl/oslbsdf.cpp
@@ -39,8 +39,8 @@
 #include "renderer/modeling/bsdf/bsdffactoryregistrar.h"
 #include "renderer/modeling/bsdf/bsdfwrapper.h"
 #include "renderer/modeling/bsdf/ibsdffactory.h"
-#include "renderer/modeling/bsdf/oslmicrofacetbrdf.h"
-#include "renderer/modeling/bsdf/oslmicrofacetbtdf.h"
+#include "renderer/modeling/bsdf/osl/oslmicrofacetbrdf.h"
+#include "renderer/modeling/bsdf/osl/oslmicrofacetbtdf.h"
 #include "renderer/modeling/bsdf/velvetbrdf.h"
 #include "renderer/modeling/input/inputevaluator.h"
 #include "renderer/modeling/scene/assembly.h"
