+++ src/appleseed/renderer/modeling/material/oslmaterial.cpp
@@ -31,7 +31,7 @@
 
 // appleseed.renderer headers.
 #include "renderer/modeling/bsdf/bsdf.h"
-#include "renderer/modeling/bsdf/oslbsdf.h"
+#include "renderer/modeling/bsdf/osl/oslbsdf.h"
 #include "renderer/modeling/bssrdf/bssrdf.h"
 #include "renderer/modeling/bssrdf/oslbssrdf.h"
 #include "renderer/modeling/edf/edf.h"
